x86 Support

+++ source/callbacks/validators/count_validator.c
@@ -1,3 +1,4 @@
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -13,17 +14,18 @@ int count_validator(argus_t *argus, void *option_ptr, validator_data_t data)
         return ARGUS_ERROR_INVALID_RANGE;
     }
     if (data.range.min > data.range.max) {
-        ARGUS_PARSING_ERROR(argus, ARGUS_ERROR_INVALID_RANGE, "Range is invalid %lld-%lld",
-                            data.range.min, data.range.max);
+        ARGUS_PARSING_ERROR(argus, ARGUS_ERROR_INVALID_RANGE,
+                            "Range is invalid %" PRId64 "-%" PRId64, data.range.min,
+                            data.range.max);
         return ARGUS_ERROR_INVALID_RANGE;
     }
 
-    long long count = option->value_count;
+    int64_t count = option->value_count;
 
     if (count < data.range.min || count > data.range.max) {
         ARGUS_PARSING_ERROR(argus, ARGUS_ERROR_INVALID_RANGE,
-                            "Values count %lld is out of range %lld-%lld", count, data.range.min,
-                            data.range.max);
+                            "Values count %" PRId64 " is out of range %" PRId64 "-%" PRId64, count,
+                            data.range.min, data.range.max);
         return ARGUS_ERROR_INVALID_RANGE;
     }
     return (ARGUS_SUCCESS);
@@ -35,6 +37,6 @@ char *format_count_validator(validator_data_t data)
     if (!result)
         return NULL;
 
-    snprintf(result, 32, "%lld-%lld", data.range.min, data.range.max);
+    snprintf(result, 32, "%" PRId64 "-%" PRId64, data.range.min, data.range.max);
     return result;
 }
